🩹 Ensure case studies can be run with non development versions (#1)
This PR ensures that the case studies can be run with the released versions of `pyglotaran` and `pyglotaran-extras`.
Main issue was that the notebooks were using a local development branch of `pyglotaran-extras` which has extra function arguments that are not present in the released version, causing crashes.

In addition it adds a `requirements.txt` with version locked dependencies to ensure that the notebooks can be run in the future even if a dependency has a breaking change.

As well as tooling to format files and run tests in the CI.
Now when making a release the CI will also build PDFs and upload them together with the rendered notebooks for easy checking of results without the need to locally run them.